transactions

+++ db/migration/000001_init_schema.up.sql
@@ -15,7 +15,7 @@ CREATE TABLE "entries" (
 
 CREATE TABLE "transfers" (
   "id" bigserial PRIMARY KEY,
-  "from_account_id" int NOT NULL,
+  "from_account_id" bigint NOT NULL,
   "to_account_id" bigint NOT NULL,
   "amount" bigint NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT (now())
