+++ db/sqlc/models.go
@@ -24,7 +24,7 @@ type Entry struct {
 
 type Transfer struct {
 	ID            int64 `json:"id"`
-	FromAccountID int32 `json:"from_account_id"`
+	FromAccountID int64 `json:"from_account_id"`
 	ToAccountID   int64 `json:"to_account_id"`
 	// must be positive
 	Amount    int64     `json:"amount"`
