+++ db/sqlc/transfer.sql.go
@@ -18,7 +18,7 @@ INSERT INTO transfers (
 `
 
 type CreateTransferParams struct {
-	FromAccountID int32 `json:"from_account_id"`
+	FromAccountID int64 `json:"from_account_id"`
 	ToAccountID   int64 `json:"to_account_id"`
 	Amount        int64 `json:"amount"`
 }
@@ -65,7 +65,7 @@ OFFSET $4
 `
 
 type ListTransfersParams struct {
-	FromAccountID int32 `json:"from_account_id"`
+	FromAccountID int64 `json:"from_account_id"`
 	ToAccountID   int64 `json:"to_account_id"`
 	Limit         int32 `json:"limit"`
 	Offset        int32 `json:"offset"`
